backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
langgraph
langchain
langchain-openai
langchain-community
faiss-cpu
pymongo
motor
httpx
numpy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
tenacity
structlog
rich
pytest
